requirements.txt
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
uwsgi